Code Generation - Simplify completion pattern (#90)
We change the pattern-matcher generated on the `completion`
parameter, to use a type ascription. This allows us to simplify
the code of the guard and the body, avoiding the use of the
`asInstanceOf` statements.

Sadly, we still need to wrap the ascription in a `Bind`

Updates to fix uncurrying args